Refactored calculator logic into a seperate class 'calcBasic'. Added tests that *only* test the happy case of each of the functions
